+++ src/main/java/com/gonzaloalonzo/spring6/repository/ILiquidacionRepo.java
@@ -3,5 +3,5 @@ package com.gonzaloalonzo.spring6.repository;
 import com.gonzaloalonzo.spring6.entity.Liquidacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ILiquidacionRepo extends JpaRepository<Liquidacion, Integer>{
+public interface ILiquidacionRepo extends JpaRepository<Liquidacion, Long>{
 }
